content/svg/content/test/test_SVGUnitTypes.html

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=366697
-->
<head>
<title>Test for Bug 842201</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=842201">Mozilla Bug 842201</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<pre id="test">
<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();

function runTest()
{
  ok(document.createElementNS("http://www.w3.org/2000/svg", "pattern")
       instanceof SVGUnitTypes, "Pattern should implement SVGUnitTypes");
  ok(document.createElementNS("http://www.w3.org/2000/svg", "filter")
       instanceof SVGUnitTypes, "Filter should implement SVGUnitTypes");
  ok(document.createElementNS("http://www.w3.org/2000/svg", "mask")
       instanceof SVGUnitTypes, "Mask should implement SVGUnitTypes");
  ok(document.createElementNS("http://www.w3.org/2000/svg", "clipPath")
       instanceof SVGUnitTypes, "ClipPath should implement SVGUnitTypes");
  ok(document.createElementNS("http://www.w3.org/2000/svg", "linearGradient")
       instanceof SVGUnitTypes, "LinearGradient should implement SVGUnitTypes");
  ok(document.createElementNS("http://www.w3.org/2000/svg", "radialGradient")
       instanceof SVGUnitTypes, "radialGradient should implement SVGUnitTypes");
  ok(!(document.createElementNS("http://www.w3.org/2000/svg", "svg")
       instanceof SVGUnitTypes), "svg should not implement SVGUnitTypes");
  SimpleTest.finish();
}

window.addEventListener("load", runTest, false);
</script>
</pre>
</body>
</html>
